requirements.txt
django-simple-captcha
sorl-thumbnail
django_tables2
django-extensions
django
django-file-resubmit
pyyaml
ua-parser
user-agents